fix(computation): use exact match in cpu lookup consumption profile function
In some cases, fuzzy match returns more than a single row. In this case, an exact match is run
to try to retrieve the correct consumption profile. It also align this comparison with the rest of the project,
which uses widely exact match. In addition, only one function depends on the look_up_consumption method, and it
expects a single value, not several.

Fix #397